streamlit_app/requirements.txt
# Core Streamlit and Web Framework
streamlit>=1.28.0
streamlit-option-menu>=0.3.6
streamlit-aggrid>=0.3.4
streamlit-ace>=0.1.1

# Azure Services
azure-storage-blob>=12.19.0
azure-identity>=1.15.0
azure-ai-documentintelligence>=1.0.0b4
azure-search-documents>=11.4.0
azure-cognitiveservices-search-websearch>=2.0.0

# OpenAI and AI
openai>=1.3.0
tiktoken>=0.5.1

# Database
psycopg2-binary>=2.9.7
sqlalchemy>=2.0.23
alembic>=1.12.1

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
openpyxl>=3.1.2
python-docx>=0.8.11
PyPDF2>=3.0.1
pdfplumber>=0.10.0
pdf2image>=1.16.3
pytesseract>=0.3.10

# Vector and Search
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
scikit-learn>=1.3.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.4.0
pydantic-settings>=2.0.3
httpx>=0.25.0
aiofiles>=23.2.0
asyncio-mqtt>=0.13.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.9.0
flake8>=6.1.0
mypy>=1.5.0

# Visualization
plotly>=5.17.0
altair>=5.1.0
matplotlib>=3.7.0
seaborn>=0.12.0
